Prepare DameCancha backend for PostgreSQL Neon and production

+++ src/disponibilidad/entities/disponibilidad.entity.ts
@@ -7,7 +7,7 @@ export class Disponibilidad {
   @PrimaryGeneratedColumn({ name: 'id_disponibilidad' })
   id_disponibilidad!: number;
 
-  @Column({ name: 'dia_semana', type: 'tinyint' })
+  @Column({ name: 'dia_semana', type: 'smallint' })
   dia_semana!: number;
 
   @Column({ name: 'hora_inicio', type: 'time' })
